add -a option for downloading or installing all translations

+++ common/fedpkg-install.sh
@@ -24,12 +24,13 @@ PROJECT_NAME=
 INPUT_FILE=
 
 LANG_CODE=
+ALL_LANGS=
 
 function usage {
     echo "This script downloads the translation of ${PROJECT_NAME}"
     echo "    usage : $0 [ARGS]"
     echo -ne "\nMandatory arguments:\n"
-    echo "   -l|--lang=LANG_CODE   Locale to pull from the server"
+    echo "   -l|--lang=LANG_CODE   Locale to pull from the server (-a : all locales)"
     echo "   -p|--project=PROJECT  Base PROJECT folder for downloaded files"
     echo "   -f|--file=INPUT_FILE  INPUT_FILE that contains the project info"
     echo "   -w|--workpath=W_PATH  Work PATH folder"
@@ -59,6 +60,11 @@ function install {
     while read -r p; do
         set -- $p
         PROJECT=${1}
+
+	if [ ! -d "${BASE_PATH}/${PROJECT}/po" ]; then
+            echo "${PROJECT}: check downloaded translations ${RED}[ FAIL ]${NC} folder not found."
+            continue
+        fi
 
         if [ -d "${BASE_PATH_RPM}/${PROJECT}" ]; then
             rm -fr "${BASE_PATH_RPM}/${PROJECT}"
@@ -113,8 +119,11 @@ function install {
         fi
 
         echo -ne "${PROJECT}: copy trans "
-        cp "${BASE_PATH}/${PROJECT}/po/${LANG_CODE}.po" "${SRC#*/}p/po/" &> /dev/null
-#        echo "cp -fu ${BASE_PATH}/${PROJECT}/po/*.po ${BASE_PATH_RPM}/${PROJECT}/${VERSION}/${SRC#*/}p/po/" | sh &> /dev/null
+        if [ -n "${ALL_LANGS}" ]; then
+            echo "cp -fu ${BASE_PATH}/${PROJECT}/po/*.po ${BASE_PATH_RPM}/${PROJECT}/${VERSION}/${SRC#*/}p/po/" | sh &> /dev/null
+        else
+            cp "${BASE_PATH}/${PROJECT}/po/${LANG_CODE}.po" "${SRC#*/}p/po/" &> /dev/null
+        fi
         if [ $? -ne 0 ]; then
             echo "${RED}[ FAIL ]${NC}"
             continue
@@ -125,7 +134,11 @@ function install {
         cd "${SRC#*/}p/po"
         intltool-update --pot
         intltool-update --dist ${LANG_CODE}
-#        for FILE in $(ls *.po); do intltool-update --dist $(basename $FILE .po); done
+        if [ -n "${ALL_LANGS}" ]; then
+            for FILE in $(ls *.po); do intltool-update --dist $(basename $FILE .po); done
+        else
+            intltool-update --dist ${LANG_CODE}
+        fi
         cd "${BASE_PATH_RPM}/${PROJECT}/${VERSION}"
 
         echo -ne "${PROJECT}: patch "
@@ -182,6 +195,9 @@ case $i in
     LANG_CODE="${i#*=}"
     shift # past argument=value
     ;;
+    -a)
+    ALL_LANGS="YES"
+    ;;
     -p=*|--project=*)
     PROJECT_NAME="${i#*=}"
     shift # past argument=value
@@ -205,7 +221,17 @@ case $i in
 esac
 done
 
-if [ -z "${LANG_CODE}" ] || [ -z "${INPUT_FILE}" ] || [ -z "${PROJECT_NAME}" ] || [ -z "${WORK_PATH}" ]; then
+if [ -z "${INPUT_FILE}" ] || [ -z "${PROJECT_NAME}" ] || [ -z "${WORK_PATH}" ]; then
+    usage
+    exit 1
+fi
+
+if [ -z "${LANG_CODE}" ] && [ -z "${ALL_LANGS}" ]; then
+    usage
+    exit 1
+fi
+
+if [ -n "${LANG_CODE}" ] && [ -n "${ALL_LANGS}" ]; then
     usage
     exit 1
 fi
@@ -218,9 +244,12 @@ VERSION=$(${WORK_PATH}/common/fedora-version.sh)
 ### Main
 # ensure running as root
 if [ "$(id -u)" != "0" ]; then
-  cd "${WORK_PATH}"
-  exec sudo "$0" "-l=${LANG_CODE}" "-p=${PROJECT_NAME}" "-f=${LIST}" "-w=${WORK_PATH}"
-  exit 0
+    if [ -n "${ALL_LANGS}" ]; then
+        exec sudo "$0" "-a" "-p=${PROJECT_NAME}" "-f=${LIST}" "-w=${WORK_PATH}"
+    else
+        exec sudo "$0" "-l=${LANG_CODE}" "-p=${PROJECT_NAME}" "-f=${LIST}" "-w=${WORK_PATH}"
+    fi
+    exit 0
 else
-  install
+    install
 fi
